requirements.txt
google-generativeai
python-decouple
fastapi[standard]
jinja2
SQLAlchemy==2.0.36
argon2-cffi==23.1.0
streamlit